requirements.txt
Django==3.0.7
django-mptt==0.10.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-cors-headers
gunicorn
martor
pillow
pylint==2.1.1